+++ ActionsScripts/Community IP Lookup.py
@@ -1,5 +1,5 @@
 import requests
-from constants import CODE_MESSAGES, USER_AGENT
+from constants import USER_AGENT
 from ScriptResult import EXECUTION_STATE_COMPLETED, EXECUTION_STATE_FAILED
 from SiemplifyAction import SiemplifyAction
 from SiemplifyDataModel import EntityTypes
@@ -7,7 +7,7 @@ from SiemplifyUtils import convert_dict_to_json_result_dict, output_handler
 
 INTEGRATION_NAME = "GreyNoise"
 
-SCRIPT_NAME = "Context IP Lookup"
+SCRIPT_NAME = "Community IP Lookup"
 
 
 @output_handler
@@ -23,7 +23,7 @@ def main():
         "Accept": "application/json",
         "Content-Type": "application/json",
         "key": api_key,
-        "User-Agent": USER_AGENT,
+        "User-Agent": "community-" + USER_AGENT,
     }
 
     ips = [
@@ -36,18 +36,14 @@ def main():
     result_value = True
     status = EXECUTION_STATE_COMPLETED
     output_json = {}
-    quick_url = "https://api.greynoise.io/v2/noise/quick/"
-    context_url = "https://api.greynoise.io/v2/noise/context/"
+    community_url = "https://api.greynoise.io/v3/community/"
     for ipaddr in ips:
         siemplify.LOGGER.info("Started processing IP: {}".format(ipaddr))
-        url = f"{quick_url}{ipaddr}"
+        url = f"{community_url}{ipaddr}"
 
         res = requests.get(url, headers=headers)
 
         if res.status_code == 200 and res.json()["noise"]:
-
-            url = f"{context_url}{ipaddr}"
-            res = requests.get(url, headers=headers)
 
             siemplify.result.add_json(str(ipaddr), res.json())
 
@@ -56,20 +52,33 @@ def main():
             output_json[str(ipaddr)] = output
 
             siemplify.add_entity_insight(
-                ipaddr, to_insight(output), triggered_by=INTEGRATION_NAME
+                ipaddr, to_noise_insight(output), triggered_by=INTEGRATION_NAME
+            )
+
+            output_message = output_message + "{},".format(ipaddr)
+
+        elif res.status_code == 200 and res.json()["riot"]:
+
+            siemplify.result.add_json(str(ipaddr), res.json())
+
+            output = res.json()
+
+            output_json[str(ipaddr)] = output
+
+            siemplify.add_entity_insight(
+                ipaddr, to_riot_insight(output), triggered_by=INTEGRATION_NAME
             )
 
             output_message = output_message + "{},".format(ipaddr)
 
         elif res.status_code == 401:
-            output_message = "Unable to auth, please check API Key.  This action requires a Paid Subscription."
+            output_message = "Unable to auth, please check API Key"
             result_value = False
             status = EXECUTION_STATE_FAILED
             siemplify.end(output_message, result_value, status)
 
         else:
             output = res.json()
-            output["message"] = CODE_MESSAGES[output["code"]]
             siemplify.result.add_json(str(ipaddr), output)
 
             output_json[str(ipaddr)] = output
@@ -82,12 +91,12 @@ def main():
     siemplify.end(output_message, result_value, status)
 
 
-def to_insight(self):
+def to_noise_insight(self):
     content = ""
     content += "<table style='100%'><tbody>"
     content += (
         "<tr><td style='text-align: left; width: 30%;'><strong style='font-size: 17px'>"
-        "Noise: <span>{noise}</span></strong></td>".format(noise=self["seen"])
+        "Noise: <span>{noise}</span></strong></td>".format(noise=self["noise"])
     )
     content += "</tbody></table><br>"
     content += (
@@ -116,27 +125,57 @@ def to_insight(self):
             "</td></tr>".format(classification=self["classification"])
         )
     content += (
+        "<tr><td style='text-align: left; width: 30%;'><strong>Name: </strong></td>"
+        "<td style='text-align: left; width: 30%;'>{name}</td></tr>".format(
+            name=self["name"]
+        )
+    )
+    content += (
         "<tr><td style='text-align: left; width: 30%;'><strong>Last Seen: </strong></td>"
         "<td style='text-align: left; width: 30%;'>{last_seen}</td></tr>".format(
             last_seen=self["last_seen"]
-        )
-    )
-    content += (
-        "<tr><td style='text-align: left; width: 30%;'><strong>Organization: </strong>"
-        "</td><td style='text-align: left; width: 30%;'>{organization}</td></tr>".format(
-            organization=self["metadata"]["organization"]
-        )
-    )
-    content += (
-        "<tr><td style='text-align: left; width: 30%;'><strong>Country: </strong></td>"
-        "<td style='text-align: left; width: 30%;'>{organization}</td></tr>".format(
-            organization=self["metadata"]["country"]
         )
     )
     content += "</tbody></table><br><br>"
     content += (
         '<p><strong>More Info: <a target="_blank" href=https://viz.greynoise.io/ip/'
         "{ip}>https://viz.greynoise.io/ip/{ip}</a></strong>&nbsp; </p>".format(
+            ip=self["ip"]
+        )
+    )
+
+    return content
+
+
+def to_riot_insight(self):
+    content = ""
+    content += "<table style='100%'><tbody>"
+    content += (
+        "<tr><td style='text-align: left; width: 30%;'><strong style='font-size: 17px;"
+        "color:#7CFC00'><span>Benign Service</span></strong></td>"
+    )
+    content += "</tbody></table><br>"
+    content += (
+        "<p>This IP is from a known harmless services and/or organizations and can "
+        "most likely be trusted.</p></br>"
+    )
+    content += "<table style='100%'><tbody>"
+    content += (
+        "<tr><td style='text-align: left; width: 30%;'><strong>Name: </strong></td>"
+        "<td style='text-align: left; width: 30%;'>{name}</td></tr>".format(
+            name=self["name"]
+        )
+    )
+    content += (
+        "<tr><td style='text-align: left; width: 30%;'><strong>Last Updated: </strong>"
+        "</td><td style='text-align: left; width: 30%;'>{last_updated}</td></tr>".format(
+            last_updated=self["last_seen"]
+        )
+    )
+    content += "</tbody></table><br><br>"
+    content += (
+        '<p><strong>More Info: <a target="_blank" href=https://viz.greynoise.io/riot/'
+        "{ip}>https://viz.greynoise.io/riot/{ip}</a></strong>&nbsp; </p>".format(
             ip=self["ip"]
         )
     )