+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.0.0
+## 2.0
+  - Fixed incorrect visualizer link in Context IP Lookup Command
+  - Added support for new Quick Lookup endpoint / codes
+  - Added new Action for Community IP Lookup, which allows for IP lookups via the free Community API
+
+## 1.0
   - Initial Release
   - Created Actions: Quick IP Lookup, Context IP Lookup, Riot IP Lookup, Execute GNQL Query
   - Created Connector: Generate Alert from GreyNoise GNQL